requirements.txt
autopep8==1.3.2
backports.weakref==1.0rc1
bleach==1.5.0
cycler==0.10.0
decorator==4.1.2
html5lib==0.9999999
Markdown==2.6.8
matplotlib==2.0.2
ml-metrics==0.1.4
networkx==1.11
numpy==1.13.0+mkl
olefile==0.44
pandas==0.20.3
Pillow==4.2.1
progressbar2==3.33.2
protobuf==3.3.0
pycodestyle==2.3.1
pyparsing==2.2.0
python-dateutil==2.6.1
python-utils==2.2.0
pytz==2017.2
PyWavelets==0.5.2
scikit-image==0.13.0
scipy==0.19.1
six==1.10.0
tensorflow==1.2.1
virtualenv==15.1.0
Werkzeug==0.12.2
